Fix #11 Incompatiblity with Search plughin
The page counting part was running at siteBodyBegin() to allow the Hitcounter displayed on the page to be correct from the start, however there seems to be some incompatability.
Changing the running order back to siteBodyEnd() fixes it so for the sake of being 1 count behind I'll leave it running at the end.